requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.0
flask==2.3.0
werkzeug==2.3.0
gunicorn==21.2.0